Set priorities appropriately for current-tripleo
There's a fair amount of complexity in the interactions of the
current-tripleo repos, so we need to fiddle the priorities of the
different repos to make sure things get installed from the correct
place.

This also means that we need to pull all the repos in the
current-tripleo and can't allow them to pulled separately or a
subsequent download may overwrite our modified repos.