Add current-tripleo-dev repo type
It is perfectly fine to use to use current-tripleo standalone, so
we shouldn't assume when someone asks for current-tripleo that they
want the whole complex dev setup.  Instead, current-tripleo should
download just the current-tripleo repo, and current-tripleo-dev
will be used to indicate that the full mixed repo config is desired.