Stop setting priority on deps repo
The deps repo is no longer generated with a priority, which makes
the sed in tripleo.sh a noop.  Attempting to set a priority the
way we do actually causes the priority of rdo-qemu-ev to be
increased, which causes problems.  Not setting any priority matches
the tripleo.sh behavior and makes things work as intended.