Update ceph and opstools repos
The ceph and opstools repos are no longer installed in the same way
they were when this was originally written.  Now it appears we are
creating repo files from scratch with the appropriate paths.

This change also begins adding support for AFS mirrors in ci by
allowing override of the centos mirror.

Change-Id: Ia006f6dc570f5041b9ed0cb3c97b82be094c0787
Partial-Bug: 1716176