Implement Proxy Support
In CI we need the ability to point at the AFS mirrors for our repo
configuration.  This change allows the ability to override the
server that the repos will be retrieved from, and rewrites the
baseurl entries in the repos to match the overridden value.  This
is because the AFS mirror repo files still point at the regular
trunk.rdoproject.org which is not what we want.

Change-Id: I0f89dd762344ab8264c1694b028e7f9e4dffb1c6
Closes-Bug: 1716176