Include python*-tripleoclient from delorean.repo
Now that tripleoclient is packaged as python2-tripleoclient for python
v2, we need to make sure we are including the latest version by using a
wildcard in the package name. Otherwise, the version of
python2-tripleoclient that ends up getting installed could be too old to
work with other versions of tripleo packages from latest delorean.

Change-Id: Ie3138801eb2e9f1c07afacbbc48d8fb3a937bcac
Closes-Bug: #1820111