Default to Ceph Nautilus unless branch is for an older OpenStack
With the release of Stein the tested version of Ceph is Nautilus.
Update tripleo-repos command to enable the Nautilus repository
by default unless an older branch is used. Add condition to set
the branch to Luminous if Queens or Rocky are used. Conditions
already exist to use the appropriate Ceph for earlier OpenStack
branches.

Also update test-requirements to get doc jobs passing,
as sphinx 2.0.0 is not compatible with python2.7.

Co-Authored-By: Alan Bishop <abishop@redhat.com>
Change-Id: Icd38d52dc5353e075d7c3bbd76169449ae043828