Fix component handling
Since the we split the repos into components, a single delorean repo
file can contain many repos. For current-tripleo-dev we only want to
include certain packages from all the current repos. This change
processes the repo file to append includepkgs to all repos defined in a
single file. Additionally it adjusts the repo names to prevent conflicts
and addresses priority is missing from a file with multiple repos.

Change-Id: Ifb301609bf29a6a171c284647baea233a7dd07cc
Closes-Bug: #1884129