Adds tripleo-get-hash module get tripleo-ci hash info from tag
This is cherrypicked from [1] where it originally merged, except
minor update for setup.cfg to point to tripleo-repos instead of
tripleo-ci and tox.ini to include tripleo-get-hash in the tox
python test discovery and execution.

We want to replace the current ansible role and bash scripts
that fetch the hash for us in tripleo CI jobs.
This tripleo-get-hash module will be packaged to pypi, then used in
an ansible python module to replace the current ansible invocations
of get-hash.

[1] https://review.opendev.org/c/openstack/tripleo-ci/+/784392/

Change-Id: I256175f55a783fe5f4e787bcb0af76bbf09cc465