Don't always validate distro name
When the output path is not /etc/yum.repos.d, the repo files won't be
used to install packages on this host, so there is no need to be
strict about host OS name validation.

This is useful when building CentOS images from a Fedora workstation.
If the host /etc/yum.repos.d is not used, the host distro is not
actually relevant.

Change-Id: I02b7a2a0d5dc6849fa04e298bce18657a9213621