Allow config overrides for TripleOHashInfo creation
Allows passing in a config dictionary into the TripleOHashInfo
constructor. This allows to override the loaded config for example
to pass in a dlrn_url. The passed config doesn't need to contain
all the config keys and the rest of the values are taken from the
loaded config file as before. This also adds a new parameter for
the ansible module so we can pass dlrn_url via ansible invocations.
Includes 2 new test cases for the new functionality. Used by [1].

[1] https://review.opendev.org/c/openstack/tripleo-quickstart/+/791486

Change-Id: I6e3e8663b7ca1ec2758843795142f91aa29fe74d