Adds python2 compatible http_get for centos7 train jobs
This adjusts the http_get function definition to work with python2.
In python3 we work with urllib.request.urlopen but for py2 we must
use urllib2.urlopen. This aims to fix the issue seen in [1] and is
needed by [2]. Adds the ansible-bad-import to the test ignores to
address issue seen in [3]

[1] https://logserver.rdoproject.org/92/35492/3/check/tripleo-ci-centos-7-content-provider/d906f0e/job-output.txt
[2] https://review.opendev.org/c/openstack/tripleo-quickstart/+/791486
[3] https://47bdf347398a802ddc78-6d4960ba8e43184f8d8ec59d1a3f8e83.ssl.cf2.rackcdn.com/809159/1/check/tox-sanity-py27/f45ffb7/job-output.txt

Change-Id: Ieb178fd9fb61284dbedea897a4841e3459e85300