Fix tripleo-get-hash tripleo-yum-config python installation
We are now consuming this as an ansible collection but it seems
we overlooked the setuptools setup.cfg install. For starters we
need the ansible module paths. Added a note to get_hash doc
about using --user for python setup.py install.
The depends-on is needed for the failing ubi-8 build job and is
not related to the change made here.

Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/814400

Change-Id: Ie5655bea08e21ce59f259832fd16cd713f90c916