yum-config compose-repos generation improvements
This patch extends the yum-config to try to update a section if it
already exists when generation compose-repos. This allows scripts to
update an existing repo from a compose-url, and also update any option
based on environment variables.
This patch also avoids compose-repos to fail when a file to be disabled
does not exists anymore.

Change-Id: I7365f7918f3ffabdd47339ad38dd97f641fe5e31
Signed-off-by: Douglas Viroel <dviroel@redhat.com>