Add RHEL/UBI 9 as a supported distros for tripleo-repos
https://review.opendev.org/c/openstack/tripleo-ansible/+/839688
moves the molecule job to use ubi9 image.

The molecule jobs are failing with missing packages as repos
are not properly configured.
Adding support for the same fixes the issue.

Bumping ansible version (which is supported in CI ) under tox.ini,
as latest version of ansible-2.13 broke the sanity job.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I71572446cc8edb9f0bce3368fb03e4f2d3488801