Use openstack-python3-zed-jobs template
Since Python 2.7 is gone EOL long time ago and C7 train is no
longer maintained, so removing it.

Python-3.6 is no longer works with Zed release so dropping it
and switching it to zed template.

From tox sanity environment, ansible-core-2.13 is getting
installed. In order to override that, we need to strictly
define it under tox sanity environment.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id9b01070b8d187b85b9efade4af548871b2ad369