Temp disable verbose logging test and py39-sanity test
As described in related-bug the tripleo-repos gate is blocked
on failing tox-cover and tox-sanity-py39 jobs.

The verbose logging test is a nice to have but not critical and we
already have py39 coverage with openstack-tox-py39. The verbose log
test can be re-enabled once we have a fix but for now we need to
unblock the gate.

Related-Bug: 2002112

Change-Id: Ia7a1e840f878def13e454c44741fb30702d8a975